Function searchEndPoint is working as intended
The function is recursive. It stores in a breadth-search style all elements with their sibling. They also store the value of their parent. When the function finds the endpoint it returns the res array. This array gets bigger every time the start(which changes at every recursive step) is equal to the previous array in the one that stores the endpoint. Then it returns entire array in reverse order